+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
+### Adde
+- **CUMULUS-556** - add mechanism for creating migration scripts
 - **CUMULUS-530** - PDR tracking through Queue-granules
   - Add optional `pdr` property to the sync-granule task's input config and output payload.
 - **CUMULUS-548** - Create a Lambda task that generates EMS distribution reports